fix: get rid of nom - moved descriptors - update to steamlocate 2 beta - move yank functions - add ARCHEAN_STEAM_ID static - update README.md - move app / match_commands to command::prelude - add some helpers for CommandError - better error handling on filesystem stuff
