+++ Cargo.toml
@@ -16,7 +16,7 @@ exclude = [
 [dependencies]
 clap = "4.5"
 clap_complete = "4.5"
-steamlocate = { version = "1.2.1", features = ["steamid_ng"] }
+steamlocate = "2.0.0-beta.2"
 serde_derive = "1.0.204"
 serde_json = "1.0.120"
 serde = { version = "1", features = ["derive"] }
