+++ src/statics.rs
@@ -1,3 +1,4 @@
 pub const COMMAND: &str = "archbelt";
 pub const VERSION: &str = "0.2.1";
 pub const DESCRIPTION: &str = "A tool to work with Archean blueprints and XenonCode modules";
+pub const ARCHEAN_STEAM_ID: u32 = 2941660;
